tested camera input
